+++ src/.internal/core/render/Text.ts
@@ -21,7 +21,7 @@ export interface ITextSettings extends ISpriteSettings {
 	fontWeight?: "normal" | "bold" | "bolder" | "lighter" | "100" | "200" | "300" | "400" | "500" | "600" | "700" | "800" | "900"
 	fontStyle?: "normal" | "italic" | "oblique";
 	fontVariant?: "normal" | "small-caps";
-	textDecoration?: "underline";
+	textDecoration?: "underline" | "line-through";
 	shadowColor?: Color | null;
 	shadowBlur?: number;
 	shadowOffsetX?: number;